Deal with singularity
To deal with the singularity of the system at x = 0, we start x at a little past 0, but leave other conditions the same. We saw this idea mentioned online somewhere, so we'll see if it works. We also saw references indicating that solving the ode is a relatively common approach to estimate Bessel functions.